requirements.txt
pytorch~=2.3.0
numpy~=1.24.4
scikit-learn~=1.5.1
pandas~=2.2.2
glob2~=0.7
matplotlib~=3.9.1
tqdm~=4.64.1